requirements.txt
pytest>=8.0.0
requests>=2.31.0
PyYAML>=6.0.1
pytest-html>=4.1.1
pytest-rerunfailures>=14.0
openpyxl>=3.1.2
jsonpath-ng>=1.6.1
